+++ lib/utils/app_location.dart
@@ -1,9 +1,7 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:location/location.dart';
 
 class AppLocation {
-
-  static Future<LocationData?> GetUserActualLocation() async {
+  static Future<LocationData?> getUserActualLocation() async {
     Location location = Location();
 
     bool _serviceEnabled = await location.serviceEnabled();
@@ -23,24 +21,5 @@ class AppLocation {
     }
 
     return await location.getLocation();
-  }
-
-  static Future UpdateUserLatLong(String? FirebaseAuthUid) async {
-    if (FirebaseAuthUid == null){
-      return;
-    }
-
-    var location = await AppLocation.GetUserActualLocation();
-    final snapshot = await FirebaseFirestore.instance
-        .collection("stores")
-        .where("firebaseAuthUid", isEqualTo: FirebaseAuthUid)
-        .limit(1)
-        .get()
-        .then((userSnapshot) =>
-        userSnapshot.docs[0].reference.update({
-          "lat": location?.latitude,
-          "long": location?.longitude
-        })
-    );
   }
 }
