+++ lib/view_models/user_register_view_model.dart
@@ -1,36 +1,15 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:grupolaranja20212/models/user.dart';
 import 'package:grupolaranja20212/models/user_type.dart';
+import 'package:grupolaranja20212/services/user_service.dart';
 
 class UserRegisterViewModel extends ChangeNotifier {
   Future<List<UserType>> getUserTypes() async {
-    var querySnapShot =
-        await FirebaseFirestore.instance.collection("user_types").get();
-
-    return querySnapShot.docs.map((e) => UserType.fromSnapshot(e)).toList();
+    return await UserService().getUserTypes();
   }
 
   Future<User?> getUserByFirebaseAuthUid(String firebaseAuthUid) async {
-    var querySnapShot = await FirebaseFirestore.instance
-        .collection("users")
-        .where('firebaseAuthUid', isEqualTo: firebaseAuthUid)
-        .get();
-
-    var foundedUsers =
-        querySnapShot.docs.map((e) => User.fromSnapshot(e)).toList();
-
-    if (foundedUsers.isEmpty) {
-      return null;
-    }
-
-    if (foundedUsers.length > 1) {
-      throw Exception(
-          "Foi encontrado mais de um usuário para o firebaseAuthUid informado: " +
-              firebaseAuthUid);
-    }
-
-    return foundedUsers[0];
+    return await UserService().getUserByFirebaseAuthUid(firebaseAuthUid);
   }
 
   Future<User> createOrUpdateUserByFirebaseAuthUid(
@@ -42,46 +21,14 @@ class UserRegisterViewModel extends ChangeNotifier {
       String description,
       int maxSearchDistance,
       List<String> skills) async {
-    var oldUser = await getUserByFirebaseAuthUid(firebaseAuthUid);
-
-    if (oldUser == null) {
-      var newUser = User(
-          firebaseAuthUid,
-          "",
-          name,
-          linkedinUrl,
-          photoUrl,
-          typeId,
-          description,
-          0,
-          0,
-          maxSearchDistance,
-          skills, <String>[], <String>[]);
-
-      await FirebaseFirestore.instance
-          .collection("stores")
-          .add(newUser.toMap());
-    } else {
-      oldUser.name = name;
-      oldUser.linkedinUrl = linkedinUrl;
-      oldUser.photoUrl = photoUrl;
-      oldUser.typeId = typeId;
-      oldUser.description = description;
-      oldUser.maxSearchDistance = maxSearchDistance;
-      oldUser.skills = skills;
-
-      await FirebaseFirestore.instance
-          .collection("stores")
-          .doc(oldUser.reference!.id)
-          .update(oldUser.toMap());
-    }
-
-    var storedUser = await getUserByFirebaseAuthUid(firebaseAuthUid);
-
-    if (storedUser != null) {
-      return storedUser;
-    }
-
-    throw Exception("Não foi possível gravar o novo usuário");
+    return await UserService().createOrUpdateUserByFirebaseAuthUid(
+        firebaseAuthUid,
+        name,
+        linkedinUrl,
+        photoUrl,
+        typeId,
+        description,
+        maxSearchDistance,
+        skills);
   }
 }
